refactor: split the post-action terminal-PR suite under the line-limit cap
test_pr_monitor_post_action_terminal_pr.py had grown to 2958 lines. It becomes
a three-part package with the shared builders in ._helpers and the PostgreSQL
factory in the package conftest, matching the outdated-resolution split. All 67
tests are preserved.

The integration runner parts get the same treatment for _pr_payload and the new
_queue_post_action_recheck, which were duplicated three ways and had pushed
part_003 over the cap.
